Near completion of new input system
Also removed some old unnecessary files.

+++ Rujin/AudioService.h
@@ -5,7 +5,7 @@
 
 namespace rujin
 {
-	using sound_id = size_t;
+	using sound_id = int32_t;
 
 	class AudioService : public Service
 	{
